requirements.txt
Flask
Flask-Cors
google-generativeai
firebase-admin
gunicorn
